+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+- Added step css handle to login content component
+
 ## [2.33.1] - 2020-06-01
 
 ### Fixed